+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for the `omitEmitField` option in the `generateCodecCode` method. When specified, schemas having a truthy value in the field indicated by the `omitEmitField` value will not be added to the type emit.
 
 ## [0.7.0-0] - 2022-05-20
 ### Added